apps/consumer-agent-cli/pyproject.toml
[project]
name = "consumer-agent-cli"
version = "0.1.0"
description = "Consumer Agent CLI for ACP Demo - uses MCP tools to discover and interact with offers"
authors = [
    {name = "ACP Demo Team", email = "demo@example.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/consumer_agent"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
